Moving match list item over.
