+++ app/src/main/java/com/adammcneilly/pocketleague/ui/EventStageDetailContent.kt
@@ -17,6 +17,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.adammcneilly.pocketleague.composables.match.MatchListItem
 import com.adammcneilly.pocketleague.shared.screens.eventstagedetail.EventStageDetailViewState
 
 /**
