+++ composables/src/commonMain/kotlin/com/adammcneilly/pocketleague/composables/match/RecentMatchCardContent.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.pocketleague.composables.recentmatch
+package com.adammcneilly.pocketleague.composables.match
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
